benches: merge bench_pipeline.cc into bench_optimize.cc
- Move pipeline fixtures & benchmarks from shared header and standalone file into bench_optimize.cc
- Delete bench_pipeline.cc
- Remove pipeline fixture definitions from bench_fixtures.hh
