+++ benches/include/bench_fixtures.hh
@@ -84,59 +84,6 @@ struct AdjacentTextParseFixture : ::benchmark::Fixture {
 };
 
 // -------------------------------------------------------------------
-// Pipeline fixtures — stores string + pre-parsed AST
-// -------------------------------------------------------------------
-struct PlainTextPipelineFixture : ::benchmark::Fixture {
-    ::fast_io::u8string input;
-    ::pltxt2htm::Ast<ndebug> ast;
-    void SetUp(::benchmark::State& state) override {
-        input = make_plain_text(as_size(state));
-        auto sv = ::fast_io::u8string_view{input.data(), input.size()};
-        ast = ::pltxt2htm::parse_pltxt<ndebug>(sv);
-    }
-};
-
-struct RedundantColorPipelineFixture : ::benchmark::Fixture {
-    ::fast_io::u8string input;
-    ::pltxt2htm::Ast<ndebug> ast;
-    void SetUp(::benchmark::State& state) override {
-        input = make_redundant_color_nesting(as_size(state));
-        auto sv = ::fast_io::u8string_view{input.data(), input.size()};
-        ast = ::pltxt2htm::parse_pltxt<ndebug>(sv);
-    }
-};
-
-struct MixedRedundantPipelineFixture : ::benchmark::Fixture {
-    ::fast_io::u8string input;
-    ::pltxt2htm::Ast<ndebug> ast;
-    void SetUp(::benchmark::State& state) override {
-        input = make_mixed_redundant(as_size(state));
-        auto sv = ::fast_io::u8string_view{input.data(), input.size()};
-        ast = ::pltxt2htm::parse_pltxt<ndebug>(sv);
-    }
-};
-
-struct AdjacentTextPipelineFixture : ::benchmark::Fixture {
-    ::fast_io::u8string input;
-    ::pltxt2htm::Ast<ndebug> ast;
-    void SetUp(::benchmark::State& state) override {
-        input = make_adjacent_text_nodes(as_size(state));
-        auto sv = ::fast_io::u8string_view{input.data(), input.size()};
-        ast = ::pltxt2htm::parse_pltxt<ndebug>(sv);
-    }
-};
-
-struct HtmlSpanAttrsPipelineFixture : ::benchmark::Fixture {
-    ::fast_io::u8string input;
-    ::pltxt2htm::Ast<ndebug> ast;
-    void SetUp(::benchmark::State& state) override {
-        input = make_html_span_attrs(as_size(state));
-        auto sv = ::fast_io::u8string_view{input.data(), input.size()};
-        ast = ::pltxt2htm::parse_pltxt<ndebug>(sv);
-    }
-};
-
-// -------------------------------------------------------------------
 // End2End fixtures — stores input string
 // -------------------------------------------------------------------
 struct FullDocE2EFixture : ::benchmark::Fixture {
